Create Foder Pages and file dashboard
