+++ react-basic-practice/src/components/index.ts
@@ -6,3 +6,4 @@ export { default as TableStatus } from "./TableStatus";
 export { default as TableFinance } from "./TableFinance";
 export { default as TableColumn } from "./TableColumn";
 export { default as OptionMenu } from "./OptionsMenu";
+export { default as Modal } from "./Modal";
