+++ react-basic-practice/src/pages/dashboard.tsx
@@ -1,0 +1,74 @@
+import React from "react";
+import { Box, ChakraProvider } from "@chakra-ui/react";
+import { theme } from "@themes";
+import { Table, TableColumn } from "@components";
+import { ICustomer } from "@types";
+import { Search, Button, Modal } from "@components";
+import { useDisclosure } from "@chakra-ui/react";
+
+const Dashboard: React.FC = () => {
+  const { isOpen, onOpen, onClose } = useDisclosure();
+  // Mock data
+  const data: ICustomer[] = [
+    {
+      name: "Ann Culhane",
+      description:
+        "Lorem ipsum dolor sit amet consectetur adipisicing elit. Voluptatibus nobis non obcaecati reiciendis pariatur. Perferendis pariatur labore, atque mollitia veniam voluptatum cum deserunt facilis at enim, fuga, provident inventore rem.",
+      status: "Open",
+      rate: "$70.00",
+      balance: "-$270.00",
+      deposit: "$500.00",
+      id: 1,
+    },
+    {
+      name: "user name",
+      description:
+        "Lorem ipsum dolor sit amet consectetur adipisicing elit. Voluptatibus nobis non obcaecati reiciendis pariatur. Perferendis pariatur labore, atque mollitia veniam voluptatum cum deserunt facilis at enim, fuga, provident inventore rem.",
+      status: "Paid",
+      rate: "$70.00",
+      balance: "$270.00",
+      deposit: "$500.00",
+      id: 2,
+    },
+    {
+      name: "user name",
+      description:
+        "Lorem ipsum dolor sit amet consectetur adipisicing elit. Voluptatibus nobis non obcaecati reiciendis pariatur. Perferendis pariatur labore, atque mollitia veniam voluptatum cum deserunt facilis at enim, fuga, provident inventore rem.",
+      status: "Due",
+      rate: "$70.00",
+      balance: "-$270.00",
+      deposit: "$500.00",
+      id: 3,
+    },
+    {
+      name: "user name",
+      description:
+        "Lorem ipsum dolor sit amet consectetur adipisicing elit. Voluptatibus nobis non obcaecati reiciendis pariatur. Perferendis pariatur labore, atque mollitia veniam voluptatum cum deserunt facilis at enim, fuga, provident inventore rem.",
+      status: "Inactive",
+      rate: "$70.00",
+      balance: "$270.00",
+      deposit: "$500.00",
+      id: 4,
+    },
+  ];
+
+  return (
+    <ChakraProvider theme={theme}>
+        <Box
+          display="flex"
+          justifyContent="space-between"
+          width="100%"
+          position="absolute"
+          padding="16px 20px"
+        >
+          <Search />
+          <Button label="+ Add Customer" variant="shadow" onClick={onOpen} />
+        </Box>
+        {/* Table demo */}
+        <Table columns={TableColumn} data={data} />
+        <Modal isOpen={isOpen} onClose={onClose}></Modal>
+    </ChakraProvider>
+  );
+};
+
+export default Dashboard;
